612: Fix: plugin 'skeleton' populates wrong skeleton file.
Skeleton tries to find a skeleton file under project dir
is a file is not explicitly configured in setup function.

If the new file is located in a different directory from
current working directory, Skeleton may treat current dir
as the project dir, which can fixed by prioritizing the path
in the autocmd table.