988: Fix: LSP default config: do not set range when 'LspFormat' is not run with range.
vim.lsp.buf.format() treats formatting with range and without range
differently, even thought the range is the whole buffer.

When 'range' is set in the options passed to vim.lsp.buf.format(), nvim
will try to find the first LS that supports
'textDocument/rangeFormatting' while it will try to find the first LS
that supports 'textDocument/formatting' when 'range' is not set in the
options.

This leads to unexpected behavior when the default range is set to the
whole buffer and 'range' is always set when calling
vim.lsp.buf.format() in user-command 'LspFormat'. When user executes
'LspFormat' without range, we expect the first LS that supports
'textDocument/formatting' to be used, but nvim will select the first LS
that supports 'textDocument/rangeFormatting' because we set the 'range'
field even when we are not using a range to execute the command.